tests(ticdc): move integration into tests (#3722)
