+++ .github/workflows/ticdc_integration.yaml
@@ -58,14 +58,14 @@ jobs:
 
       - name: Build Integration Framework
         run: |
-          cd $GITHUB_WORKSPACE/integration
+          cd $GITHUB_WORKSPACE/tests/mq_protocol_tests
           go build
 
       - name: Run Integration Framework
         timeout-minutes: 45
         run: |
-          cd $GITHUB_WORKSPACE/integration
-          ./integration -protocol=mysql
+          cd $GITHUB_WORKSPACE/tests/mq_protocol_tests
+          ./mq_protocol_tests -protocol=mysql
 
       - name: Try to dump container logs
         if: ${{ failure() }}
@@ -120,14 +120,14 @@ jobs:
 
       - name: Build Integration Framework
         run: |
-          cd $GITHUB_WORKSPACE/integration
+          cd $GITHUB_WORKSPACE/tests/mq_protocol_tests
           go build
 
       - name: Run Integration Framework
         timeout-minutes: 45
         run: |
-          cd $GITHUB_WORKSPACE/integration
-          ./integration -protocol=simple-mysql-checking-old-value
+          cd $GITHUB_WORKSPACE/tests/mq_protocol_tests
+          ./mq_protocol_tests -protocol=simple-mysql-checking-old-value
 
       - name: Try to dump container logs
         if: ${{ failure() }}
@@ -182,14 +182,14 @@ jobs:
 
       - name: Build Integration Framework
         run: |
-          cd $GITHUB_WORKSPACE/integration
+          cd $GITHUB_WORKSPACE/tests/mq_protocol_tests
           go build
 
       - name: Run Integration Framework
         timeout-minutes: 45
         run: |
-          cd $GITHUB_WORKSPACE/integration
-          ./integration -protocol=canalJson
+          cd $GITHUB_WORKSPACE/tests/mq_protocol_tests
+          ./mq_protocol_tests -protocol=canalJson
 
       - name: Try to dump container logs
         if: ${{ failure() }}
