+++ tests/mq_protocol_tests/cases/case_simple.go
@@ -11,33 +11,33 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package tests
+package cases
 
 import (
 	"github.com/pingcap/errors"
-	"github.com/pingcap/ticdc/integration/framework"
+	"github.com/pingcap/ticdc/tests/mq_protocol_tests/framework"
 )
 
-// CompositePKeyCase is base impl of test case for composite primary keys
-type CompositePKeyCase struct {
+// SimpleCase is base impl of simple test case
+type SimpleCase struct {
 	framework.Task
 }
 
-// NewCompositePKeyCase create a test case which have composite primary key
-func NewCompositePKeyCase(task framework.Task) *CompositePKeyCase {
-	return &CompositePKeyCase{
+// NewSimpleCase create a test case which has some simple dmls, ddls
+func NewSimpleCase(task framework.Task) *SimpleCase {
+	return &SimpleCase{
 		Task: task,
 	}
 }
 
 // Name impl framework.Task interface
-func (s *CompositePKeyCase) Name() string {
-	return "Composite Primary Key"
+func (s *SimpleCase) Name() string {
+	return "Simple"
 }
 
 // Run impl framework.Task interface
-func (s *CompositePKeyCase) Run(ctx *framework.TaskContext) error {
-	_, err := ctx.Upstream.ExecContext(ctx.Ctx, "create table test (id1 int, id2 int, value int, primary key (id1, id2))")
+func (s *SimpleCase) Run(ctx *framework.TaskContext) error {
+	_, err := ctx.Upstream.ExecContext(ctx.Ctx, "create table test (id int primary key, value int)")
 	if err != nil {
 		return err
 	}
@@ -46,17 +46,32 @@ func (s *CompositePKeyCase) Run(ctx *framework.TaskContext) error {
 	table := ctx.SQLHelper().GetTable("test")
 	// Create an SQL request, send it to the upstream, wait for completion and check the correctness of replication
 	err = table.Insert(map[string]interface{}{
-		"id1":   0,
-		"id2":   1,
+		"id":    0,
 		"value": 0,
 	}).Send().Wait().Check()
 	if err != nil {
 		return errors.AddStack(err)
 	}
 
+	// To wait on a batch of SQL requests, create a slice of Awaitables
+	reqs := make([]framework.Awaitable, 0)
+	for i := 1; i < 1000; i++ {
+		// Only send, do not wait
+		req := table.Insert(map[string]interface{}{
+			"id":    i,
+			"value": i,
+		}).Send()
+		reqs = append(reqs, req)
+	}
+
+	// Wait on SQL requests in batch and check the correctness
+	err = framework.All(ctx.SQLHelper(), reqs).Wait().Check()
+	if err != nil {
+		return err
+	}
+
 	err = table.Upsert(map[string]interface{}{
-		"id1":   0,
-		"id2":   1,
+		"id":    0,
 		"value": 1,
 	}).Send().Wait().Check()
 	if err != nil {
@@ -64,8 +79,7 @@ func (s *CompositePKeyCase) Run(ctx *framework.TaskContext) error {
 	}
 
 	err = table.Delete(map[string]interface{}{
-		"id1": 0,
-		"id2": 1,
+		"id": 0,
 	}).Send().Wait().Check()
 	return err
 }
