+++ tests/mq_protocol_tests/cases/case_unsigned.go
@@ -11,10 +11,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package tests
+package cases
 
 import (
-	"github.com/pingcap/ticdc/integration/framework"
+	"github.com/pingcap/ticdc/tests/mq_protocol_tests/framework"
 )
 
 // UnsignedCase is base impl of test case for unsigned int type data
