+++ tests/mq_protocol_tests/framework/canal/kafka_docker_env_test.go
@@ -11,18 +11,18 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package avro
+package canal
 
 import (
 	"os/exec"
 	"testing"
 
 	"github.com/pingcap/log"
-	"github.com/pingcap/ticdc/integration/framework"
+	"github.com/pingcap/ticdc/tests/mq_protocol_tests/framework"
 	"github.com/stretchr/testify/require"
 )
 
-func TestAvroKafkaDockerEnv_Basic(t *testing.T) {
+func TestKafkaDockerEnv_Basic(t *testing.T) {
 	env := NewKafkaDockerEnv("")
 	require.NotNil(t, env)
 
@@ -52,9 +52,10 @@ func (t *dummyTask) Prepare(taskContext *framework.TaskContext) error {
 
 func (t *dummyTask) GetCDCProfile() *framework.CDCProfile {
 	return &framework.CDCProfile{
-		PDUri:   framework.UpstreamPD,
-		SinkURI: "kafka://kafka:9092/testdb_test?protocol=avro",
-		Opts:    map[string]string{"registry": "http://schema-registry:8081"},
+		PDUri:      framework.UpstreamPD,
+		SinkURI:    "kafka://kafka:9092/testdb?protocol=canal",
+		Opts:       map[string]string{"force-handle-key-pkey": "true"},
+		ConfigFile: "/configs/canal-test-config.toml",
 	}
 }
 
@@ -76,7 +77,7 @@ func (t *dummyTask) Run(taskContext *framework.TaskContext) error {
 	return nil
 }
 
-func TestAvroKafkaDockerEnv_RunTest(t *testing.T) {
+func TestCanalKafkaDockerEnv_RunTest(t *testing.T) {
 	env := NewKafkaDockerEnv("")
 	require.NotNil(t, env)
